Refactor playback logic into ActivityPlayer
