+++ app/src/main/kotlin/com/d4rk/englishwithlidia/plus/app/lessons/details/ui/LessonActivity.kt
@@ -3,7 +3,7 @@ package com.d4rk.englishwithlidia.plus.app.lessons.details.ui
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
+import com.d4rk.englishwithlidia.plus.app.player.ActivityPlayer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
@@ -11,8 +11,8 @@ import androidx.compose.ui.Modifier
 import com.d4rk.android.libs.apptoolkit.app.theme.style.AppTheme
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class LessonActivity : AppCompatActivity() {
-    private val viewModel: LessonViewModel by viewModel()
+class LessonActivity : ActivityPlayer() {
+    override val viewModel: LessonViewModel by viewModel()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
