+++ app/src/main/kotlin/com/d4rk/englishwithlidia/plus/app/lessons/details/ui/LessonScreen.kt
@@ -38,12 +38,13 @@ fun LessonScreen(
                 NoDataScreen()
             },
             onSuccess = { lesson ->
-                LessonContentLayout(
-                    paddingValues = paddingValues,
-                    scrollState = scrollState,
-                    lesson = lesson,
-                    viewModel = viewModel,
-                )
+                    LessonContentLayout(
+                        paddingValues = paddingValues,
+                        scrollState = scrollState,
+                        lesson = lesson,
+                        activity = activity,
+                        viewModel = viewModel,
+                    )
             },
         )
     }
